requirements.txt
anthropic
yfinance
pandas
requests
python-telegram-bot
python-decouple
